+++ app/src/main/kotlin/net/primal/android/core/compose/feed/model/FeedPostStatsUi.kt
@@ -1,13 +1,13 @@
 package net.primal.android.core.compose.feed.model
 
 data class FeedPostStatsUi(
-    val repliesCount: Int = 0,
+    val repliesCount: Long = 0,
     val userReplied: Boolean = false,
-    val zapsCount: Int = 0,
-    val satsZapped: Int = 0,
+    val zapsCount: Long = 0,
+    val satsZapped: Long = 0,
     val userZapped: Boolean = false,
-    val likesCount: Int = 0,
+    val likesCount: Long = 0,
     val userLiked: Boolean = false,
-    val repostsCount: Int = 0,
+    val repostsCount: Long = 0,
     val userReposted: Boolean = false,
 )
